+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: dmendoza <dmendoza@student.42barcelon      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/19 16:00:15 by dmendoza          #+#    #+#             */
-/*   Updated: 2024/12/19 16:08:32 by dmendoza         ###   ########.fr       */
+/*   Updated: 2024/12/20 15:59:49 by dmendoza         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = -1;
 	while (s[++i])
